Move the tests out of the package directory, and split examples into those inside and those outside the package directory
