+++ pyproject.toml
@@ -44,7 +44,8 @@ ignore = ["D412", "D416", "E501", "F401", "F821", "FBT0", "N802", "PIE790", "PLR
 fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
 unfixable = []
 
-# Exclude a variety of commonly ignored directories.
+# Exclude directories which are non-code. All add 'tests' as this (deliberately)
+# breaks some of the standard rules to ensure compliance.
 exclude = [
     ".eggs",
     ".git",
@@ -56,6 +57,7 @@ exclude = [
     "__pypackages__",
     "build",
     "dist",
+    "tests",
 ]
 
 # Ignore slightly long lines (often generated by docformatter or black)
