+++ tests/test_simple_delete.py
